Refactor [Image|Video]Formatter[s] into Inserters
